🚚 move logs panel to side panel #4
