+++ src/app/side-panel/logs-panel/logs-panel.component.html
