+++ src/app/side-panel/logs-panel/logs-panel.component.scss
