+++ src/app/side-panel/logs-panel/logs-panel.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { SocketService } from '../socket.service';
+import { SocketService } from '../../socket.service';
 
 @Component({
   selector: 'app-logs-panel',
